+++ retriever/__main__.py
@@ -8,7 +8,6 @@ The main() function can be used for bootstrapping.
 """
 from __future__ import print_function
 from __future__ import absolute_import
-from builtins import str
 from builtins import input
 from imp import reload
 import os
@@ -21,13 +20,16 @@ reload(sys)
 if hasattr(sys, 'setdefaultencoding'):
     sys.setdefaultencoding(encoding)
 
-from retriever import SCRIPT_LIST, sample_script, CITATION
+from retriever.lib.defaults import sample_script, CITATION
+from retriever.lib.scripts import SCRIPT_LIST
 from retriever.engines import engine_list
 from retriever.lib.repository import check_for_updates
-from retriever.lib.tools import choose_engine, name_matches, reset_retriever
+from retriever.lib.tools import name_matches, reset_retriever
+from retriever.engines import choose_engine
 from retriever.lib.get_opts import parser
 from retriever.lib.datapackage import create_json, edit_json, delete_json, get_script_filename
 from retriever.lib.datasets import datasets
+
 
 def main():
     """This function launches the Data Retriever."""
@@ -43,7 +45,7 @@ def main():
         args = parser.parse_args()
 
         if args.command == "install" and not args.engine:
-            parser.parse_args(['install','-h'])
+            parser.parse_args(['install', '-h'])
 
         if args.quiet:
             sys.stdout = open(os.devnull, 'w')
