+++ retriever/engines/csvengine.py
@@ -3,7 +3,8 @@ from builtins import object
 import os
 
 from retriever.lib.models import Engine
-from retriever import DATA_DIR, open_fw, open_csvw
+from retriever import open_fw, open_csvw
+from retriever.lib.defaults import DATA_DIR
 from retriever.lib.tools import sort_csv
