+++ retriever/engines/xmlengine.py
@@ -2,8 +2,9 @@ import os
 
 from builtins import object
 from retriever.lib.models import Engine
-from retriever import DATA_DIR, open_fr, open_fw
+from retriever import open_fr, open_fw
 from retriever.lib.tools import xml2csv, sort_csv
+from retriever.lib.defaults import DATA_DIR
 
 
 class DummyConnection(object):
