+++ scripts/aquatic_animal_excretion.py
@@ -7,7 +7,7 @@ import os
 
 from retriever.lib.models import Table
 from retriever.lib.templates import Script
-from retriever import VERSION
+from retriever.lib.defaults import VERSION
 from pkg_resources import parse_version
  
 class main(Script):
