+++ scripts/predator_prey_size_marine.py
@@ -5,7 +5,7 @@ from pkg_resources import parse_version
 
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
 from retriever.lib.templates import Script
-from retriever import VERSION
+from retriever.lib.defaults import VERSION
 
 
 class main(Script):
